pyproject.toml
[project]
name = "personal-finance-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.5",
    "django-cors-headers>=4.6.0",
    "djangorestframework>=3.15.2",
]
